+++ lib/deserializers.php
@@ -1,0 +1,86 @@
+<?php
+
+namespace Sabre\Xml\Deserializer;
+
+use Sabre\Xml\Reader;
+
+/**
+ * This class provides a number of 'deserializer' helper functions.
+ * These can be used to easily specify custom deserializers for specific
+ * XML elements.
+ */
+
+/*
+ * The 'keyValue' deserializer parses all child elements, and outputs them as
+ * a "key=>value" array.
+ *
+ * For example, keyvalue will parse:
+ *
+ * <?xml version="1.0"?>
+ * <s:root xmlns:s="http://sabredav.org/ns">
+ *   <s:elem1>value1</s:elem1>
+ *   <s:elem2>value2</s:elem2>
+ *   <s:elem3 />
+ * </s:root>
+ *
+ * Into:
+ *
+ * [
+ *   "{http://sabredav.org/ns}elem1" => "value1",
+ *   "{http://sabredav.org/ns}elem2" => "value2",
+ *   "{http://sabredav.org/ns}elem3" => null,
+ * ];
+ *
+ * If you specify the 'namespace' argument, the deserializer will remove
+ * the namespaces of the keys that match that namespace.
+ *
+ * For example, if you call keyValue like this:
+ *
+ * keyValue($reader, 'http://sabredav.org/ns')
+ *
+ * it's output will instead be:
+ *
+ * [
+ *   "elem1" => "value1",
+ *   "elem2" => "value2",
+ *   "elem3" => null,
+ * ];
+ *
+ * Attributes will be removed from the top-level elements. If elements with
+ * the same name appear twice in the list, only the last one will be kept.
+ *
+ *
+ * @param Xml\Reader $reader
+ * @param string $namespace
+ * @return array
+ */
+function keyValue(Reader $reader, $namespace = null) {
+
+    // If there's no children, we don't do anything.
+    if ($reader->isEmptyElement) {
+        $reader->next();
+        return [];
+    }
+
+    $values = [];
+
+    $reader->read();
+    do {
+
+        if ($reader->nodeType === Reader::ELEMENT) {
+            if ($namespace !== null && $reader->namespaceURI === $namespace) {
+                $values[$reader->localName] = $reader->parseCurrentElement()['value'];
+            } else {
+                $clark = $reader->getClark();
+                $values[$clark] = $reader->parseCurrentElement()['value'];
+            }
+        } else {
+            $reader->read();
+        }
+    } while ($reader->nodeType !== Reader::END_ELEMENT);
+
+    $reader->read();
+
+    return $values;
+
+}